refactor -dns flag
